производит разделение по модулям
